Relax event-monitor/handle-event parameter spec.
Changed the `context` parameter from an Atom containing a Graph with an
associated watcher to just an Atom containing a Graph, with an
associated watcher or not.

While the function is most likely to receive monitored Graphs as
arguments in run-time, it does not need to (and it shouldn't) access any
value related to the watcher thread.

The reason I made this change is that I needed to pass a non-monitored
graph for a regression test.